Change getType() of constant, type and expression into getTypeOf Add a new getType() into Expression interface Starting to implement cast support into llvm Add a new getCast() method into Instruction interface Add llvm header inside the template (to be removed later) Starting to implement array gestion in LLVM And that's all folks

+++ plugin/src/net/sf/orcc/interpreter/ListAllocator.java
@@ -50,7 +50,7 @@ public class ListAllocator extends AbstractTypeVisitor {
 		List<Integer> sizeList = type.getDimensions();
 		int[] dimensions = new int[sizeList.size()];
 		for (int i = 0; i < sizeList.size(); i++) {
-			dimensions[i] = (Integer) sizeList.get(i);
+			dimensions[i] = sizeList.get(i);
 		}
 		sizeList.clear();
